# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Board API",
        "description": "This is the Job Board API documentation.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/company/job-postings/{id}": {
            "put": {
                "tags": [
                    "Company"
                ],
                "summary": "Update an existing job posting",
                "operationId": "4709c14ee9595385a026add9e237085f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job posting",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "location",
                                    "type",
                                    "salary_min",
                                    "salary_max",
                                    "requirements",
                                    "deadline"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Backend Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description for Backend Engineer"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Nairobi"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "full-time"
                                    },
                                    "salary_min": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 350000
                                    },
                                    "salary_max": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 450000
                                    },
                                    "requirements": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Laravel",
                                            "PHP",
                                            "PostgreSQL",
                                            "REST APIs"
                                        ]
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job updated successfully"
                                        },
                                        "job": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Backend Engineer"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-04T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "The title field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Company"
                ],
                "summary": "Delete a job posting",
                "operationId": "6024d7cbf0b43a7a4e9eb890e76a7fdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job posting",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}